+++ src/cmd/mod.rs
@@ -3,4 +3,8 @@ pub mod commit;
 pub mod find;
 pub mod log;
 pub mod rebase;
+pub mod rebase_branch;
+pub mod rebase_last;
+pub mod rebase_plan;
+pub mod rebase_range;
 pub mod status;
