+++ src/cmd/rebase.rs
@@ -413,72 +413,6 @@ fn divergence_flow(
     Ok(Some(commit_oid))
 }
 
-fn last_n_flow(repo: &GitRepo) -> anyhow::Result<Option<Oid>> {
-    let n: usize;
-
-    loop {
-        let input =
-            match print_input_prompt("Specify a valid number", None)?
-            {
-                Some(i) => i,
-                None => {
-                    println!("Exiting...");
-                    return Ok(None);
-                }
-            };
-
-        match input.parse::<usize>() {
-            Ok(v) => {
-                if v == 0 {
-                    println!("Please enter a value greater than 0");
-                    continue;
-                }
-
-                n = v;
-                break;
-            }
-            Err(_) => {
-                println!("Cannot parse {} as a valid number", input);
-                continue;
-            }
-        }
-    }
-
-    let logs =
-        get_logs(repo, false, false, n, false, None, None, None)?;
-
-    // if n exceeds log length, continue, regardless
-    if n > logs.git_logs.len() {
-        println!(
-            "Only {} commits exist in history but you requested {}",
-            style(logs.git_logs.len()).red(),
-            style(n).red()
-        );
-    }
-
-    // this should get the last logged commit
-    // if the count exceeds, get_logs()
-    // will handle that and return or "take"
-    // the last commit
-    let oldest_commit_hash = match logs
-        .git_logs
-        .last()
-        .map(|l| {
-            l.commit_hash
-                .to_owned()
-        }) {
-        Some(h) => h,
-        None => {
-            println!("No Commits Found, Exiting...");
-            return Ok(None);
-        }
-    };
-
-    let oid = find_parent_commit(&repo.repo, &oldest_commit_hash)?;
-
-    Ok(Some(oid))
-}
-
 fn specify_range_flow(repo: &GitRepo) -> anyhow::Result<Option<Oid>> {
     let logs =
         get_logs(repo, false, false, 0, false, None, None, None)?;
